// apps/server/src/playwright/type.ts
export enum BrowserEngine {
    Chromium = 'chromium',
    Firefox = 'firefox',
    Webkit = 'webkit',
}
export enum BrowserDevice {
    Mobile = 'mobile',
    Desktop = 'desktop',
}
export enum BrowseDevtool {
    Script = 'script',
    Network = 'network',
    Dom = 'dom',
    Console = 'console',
    Cookie = 'cookie', //not implement
    Global = 'global',
    Application = 'application', //not implement
}